GL_GLEXT_PROTOTYPES was declared too late in ogl_defs.h, causing some OpenGL functions not to be declared before use.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1303 01821946-851b-0410-b7f4-f1477c70e57a

+++ render/render.cpp
@@ -1956,7 +1956,7 @@ if (nClearWindow == 2) {
 	if (nFirstTerminalSeg < gameData.render.mine.nRenderSegs) {
 		int i;
 
-		if (nClearWindowColor == -1)
+		if (nClearWindowColor == (unsigned int) -1)
 			nClearWindowColor = BLACK_RGBA;
 		GrSetColor (nClearWindowColor);
 		for (i = nFirstTerminalSeg, rwP = renderWindows; i < gameData.render.mine.nRenderSegs; i++, rwP++) {
